refactor(desktop): move types.rs test module to sibling tests.rs
The inline test module pushed types.rs to 1075 lines, over the 1000-line
file-size limit enforced by check:file-sizes. Extract the #[cfg(test)] mod
into a sibling types/tests.rs, matching the existing convention used by
env_vars.rs, personas.rs, and runtime.rs in this crate. Production code is
unchanged; all 22 type tests still execute by name.
